requirements.txt
streamlit
pandas
numpy
joblib
scikit-learn
xgboost
sqlalchemy
psycopg2-binary
